+++ golang/cosmos/x/swingset/genesis.go
@@ -2,6 +2,7 @@ package swingset
 
 import (
 	// "fmt"
+	stdlog "log"
 
 	"github.com/Agoric/agoric-sdk/golang/cosmos/x/swingset/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -29,7 +30,17 @@ func InitGenesis(ctx sdk.Context, keeper Keeper, data *types.GenesisState) []abc
 		storage.Value = value
 		keeper.SetStorage(ctx, key, &storage)
 	}
-	return []abci.ValidatorUpdate{}
+
+	// Just run the SwingSet kernel to finish bootstrap and get ready to open for
+	// business.
+	stdlog.Println("Running SwingSet until bootstrap is ready")
+	bootstrapBlock := abci.RequestEndBlock{Height: ctx.BlockHeight()}
+	valUpdate, err := EndBlock(ctx, bootstrapBlock, keeper)
+	if err != nil {
+		// Errors here are fatal.
+		panic(err)
+	}
+	return valUpdate
 }
 
 func ExportGenesis(ctx sdk.Context, k Keeper) *types.GenesisState {