+++ packages/pegasus/test/fakeVatAdmin.js
@@ -1,0 +1,22 @@
+import { E } from '@agoric/eventual-send';
+import { makePromiseKit } from '@agoric/promise-kit';
+
+import { evalContractBundle } from '@agoric/zoe/src/contractFacet/evalContractCode';
+
+export default harden({
+  createVat: bundle => {
+    return harden({
+      root: E(evalContractBundle(bundle)).buildRootObject(),
+      adminNode: {
+        done: () => {
+          return makePromiseKit().promise;
+        },
+        terminate: () => {},
+        adminData: () => ({}),
+      },
+    });
+  },
+  createVatByName: _name => {
+    throw Error(`createVatByName not supported in fake mode`);
+  },
+});