+++ packages/deploy-script-support/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.2](https://github.com/Agoric/agoric-sdk/compare/@agoric/deploy-script-support@0.2.1...@agoric/deploy-script-support@0.2.2) (2021-04-13)
+
+**Note:** Version bump only for package @agoric/deploy-script-support
+
+
+
+
+
 ## [0.2.1](https://github.com/Agoric/agoric-sdk/compare/@agoric/deploy-script-support@0.2.0...@agoric/deploy-script-support@0.2.1) (2021-04-07)
 
 **Note:** Version bump only for package @agoric/deploy-script-support